// FE/src/shared/api/schema.d.ts
export interface components {
    schemas: {
// ...
        ItineraryDayResponse: {
            /** Format: uuid */
            id?: string;
            /** Format: int32 */
            dayNumber?: number;
            /** Format: date */
            date?: string;
            /** Format: int64 */
            version?: number;
            items?: components["schemas"]["ItineraryItemResponse"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}